+++ src/main/java/com/career/careerlink/notice/repository/NoticeDetailRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NoticeDetailRepository extends JpaRepository<Notice, Long> {
+public interface NoticeDetailRepository extends JpaRepository<Notice, Integer> {
 }
